\name{Internal Functions}
\alias{pmclust.reduceK}
\title{ All Internal Functions }
\description{ All internal functions }
\references{
  Programming with Big Data in R Website:
  \url{https://pbdr.org/}
}
\author{
  Wei-Chen Chen \email{wccsnow@gmail.com} and George Ostrouchov.
}
\keyword{internal}
